Half of experts store is finished
