+++ src/models/experts/selectors.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unsafe-return */
-import { RootStateType } from '../rootReducer';
+import { RootStateType } from '../../old/store/rootReducer';
 import { IExpertsState } from './types';
 
 export const selectExperts = (state: RootStateType): IExpertsState =>
