+++ src/models/materials/reducers.ts
@@ -1,9 +1,9 @@
 /* eslint-disable */
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { IPost, LoadingStatusEnum } from '../../lib/types';
+import { IPost, LoadingStatusEnum } from '../../old/lib/types';
 import { IMaterialsState } from './types';
 import { fetchMaterials } from './asyncActions';
-import { getAsyncActionsReducer } from '../helpers/asyncActions';
+import { getAsyncActionsReducer } from '../../old/store/helpers/asyncActions';
 
 const initialState: IMaterialsState = {
   data: {
