+++ src/models/materials/selectors.ts
@@ -1,4 +1,4 @@
-import { RootStateType } from '../rootReducer';
+import { RootStateType } from '../../old/store/rootReducer';
 import { IMaterialsState } from './types';
 
 export const selectMaterials = (state: RootStateType): IMaterialsState =>
