+++ src/old/modules/main/store/mainSlice.ts
@@ -2,7 +2,7 @@
 import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { LoadingStatusEnum } from '../../../lib/types';
 import { getPosts } from '../../../lib/utilities/API/api';
-import { loadPosts, mapFetchedPosts } from '../../../store/materials';
+import { loadPosts, mapFetchedPosts } from '../../../../models/materials';
 import {
   LOAD_POSTS_LIMIT,
   LOAD_IMPORTANT_POSTS_LIMIT,
